Refactored inheritance for tagged resources to composition pattern.
Fixed a few minor nits from PR.
