+++ samples/explore_datasource.py
@@ -37,6 +37,7 @@ def main():
     # SIGN IN
     tableau_auth = TSC.TableauAuth(args.username, password)
     server = TSC.Server(args.server)
+    server.use_highest_version()
     with server.auth.sign_in(tableau_auth):
         # Query projects for use when demonstrating publishing and updating
         all_projects, pagination_item = server.projects.get()
@@ -68,7 +69,6 @@ def main():
                    for connection in sample_datasource.connections])
 
             # Add some tags to the datasource
-            server.version = 2.6 # Datasource tagging requires server version 2.6
             original_tag_set = set(sample_datasource.tags)
             sample_datasource.tags.update('a', 'b', 'c', 'd')
             server.datasources.update(sample_datasource)
