+++ tableauserverclient/models/datasource_item.py
@@ -2,25 +2,25 @@ import xml.etree.ElementTree as ET
 from .exceptions import UnpopulatedPropertyError
 from .property_decorators import property_not_nullable
 from .tag_item import TagItem
-from .tagged_resource_item import TaggedResourceItem
 from .. import NAMESPACE
 from ..datetime_helpers import parse_datetime
 import copy
 
 
-class DatasourceItem(TaggedResourceItem):
+class DatasourceItem(object):
     def __init__(self, project_id, name=None):
-        super(DatasourceItem, self).__init__()  # Python2 compatible super
         self._connections = None
         self._content_url = None
         self._created_at = None
-        self._id = None
-        self._project_name = None
         self._datasource_type = None
+        self._id = None
+        self._initial_tags = set()
+        self._project_name = None
         self._updated_at = None
         self.name = name
         self.owner_id = None
         self.project_id = project_id
+        self.tags = set()
 
     @property
     def connections(self):
@@ -61,6 +61,12 @@ class DatasourceItem(TaggedResourceItem):
     @property
     def updated_at(self):
         return self._updated_at
+
+    def _get_initial_tags(self):
+        return self._initial_tags
+
+    def _set_initial_tags(self, initial_tags):
+        self._initial_tags = initial_tags
 
     def _set_connections(self, connections):
         self._connections = connections
