+++ tableauserverclient/models/workbook_item.py
@@ -2,20 +2,19 @@ import xml.etree.ElementTree as ET
 from .exceptions import UnpopulatedPropertyError
 from .property_decorators import property_not_nullable, property_is_boolean
 from .tag_item import TagItem
-from .tagged_resource_item import TaggedResourceItem
 from .view_item import ViewItem
 from .. import NAMESPACE
 from ..datetime_helpers import parse_datetime
 import copy
 
 
-class WorkbookItem(TaggedResourceItem):
+class WorkbookItem(object):
     def __init__(self, project_id, name=None, show_tabs=False):
-        super(WorkbookItem, self).__init__()  # Python2 compatible super
         self._connections = None
         self._content_url = None
         self._created_at = None
         self._id = None
+        self._initial_tags = set()
         self._preview_image = None
         self._project_name = None
         self._size = None
@@ -25,6 +24,7 @@ class WorkbookItem(TaggedResourceItem):
         self.owner_id = None
         self.project_id = project_id
         self.show_tabs = show_tabs
+        self.tags = set()
 
     @property
     def connections(self):
@@ -88,6 +88,12 @@ class WorkbookItem(TaggedResourceItem):
             error = "Workbook item must be populated with views first."
             raise UnpopulatedPropertyError(error)
         return self._views
+
+    def _get_initial_tags(self):
+        return self._initial_tags
+
+    def _set_initial_tags(self, initial_tags):
+        self._initial_tags = initial_tags
 
     def _set_connections(self, connections):
         self._connections = connections
